+++ test/test_camera.py
@@ -29,6 +29,7 @@ from utils import *
 
 class test_camera(unittest.TestCase):
     def setUp(self):
+        setUpTest()
         cmds.file(new = True, force = True)
         # Set current time to be 30fps
         cmds.currentUnit( time='ntsc' )
@@ -40,6 +41,7 @@ class test_camera(unittest.TestCase):
         for f in self.__files:
             if os.path.exists(f):
                 os.remove(f)
+        tearDownTest()
                 
     def ExportJsonData(self, objectName):
         # Export Json