+++ test/test_customAttributes.py
@@ -33,6 +33,7 @@ class test_customAttributes(unittest.TestCase):
     nodesWithCustomAttrs = 0
 
     def setUp(self):
+        setUpTest()
         cmds.file(new = True, force = True)
         # Set current time to be 30fps
         cmds.currentUnit( time='ntsc' )
@@ -44,6 +45,7 @@ class test_customAttributes(unittest.TestCase):
         for f in self.__files:
             if os.path.exists(f):
                 os.remove(f)
+        tearDownTest()
     
 
     def ExportJsonData(self):